+++ package/Runtime/LocalizatorRequest.cs
@@ -13,19 +13,20 @@ namespace AlchemyBow.Localizations
     {
         private const string ResourcesFilePath = "AlchemyLocalizations/localizations";
 
-        private readonly LocalizatorConfig config;
+        private readonly LocalizatorRequestSettings settings;
         private readonly string language;
         private readonly string[][] localizations;
-        private readonly HashSet<int> groups;
+        private readonly Dictionary<string, int> stringMap;
 
         private readonly Action successfullyCompleded;
 
-        private LocalizatorRequest(LocalizatorConfig config, string language, HashSet<int> groups,
-            string[][] localizations, Action successfullyCompleded)
+        private LocalizatorRequest(LocalizatorRequestSettings settings, string language,
+            string[][] localizations, Dictionary<string, int> stringMap, Action successfullyCompleded)
         {
-            this.config = config;
+            this.settings = settings;
             this.language = language;
             this.localizations = localizations;
+            this.stringMap = stringMap;
             this.successfullyCompleded = successfullyCompleded;
         }
 
@@ -46,17 +47,17 @@ namespace AlchemyBow.Localizations
         /// Schedules the specified language to load asynchronously and returns a handle for the process.
         /// </summary>
         /// <param name="languageIndex">An index of the language to load.</param>
-        /// <param name="groups">Indexes of the groups to load.</param>
-        /// <param name="config">A configuration of the Localizator.</param>
-        /// <param name="localizations">A load destination</param>
+        /// <param name="settings">A settings object that defines what and how to load.</param>
+        /// <param name="localizations">A load destination.</param>
+        /// <param name="stringMap">A load destination.</param>
         /// <param name="successfullyCompleded">An action to invoke when the process is successfully completed.</param>
         /// <returns>A handle for the loading process.</returns>
-        public static LocalizatorRequest RequestLocalizations(int languageIndex, HashSet<int> groups, 
-            LocalizatorConfig config, string[][] localizations, Action successfullyCompleded)
+        public static LocalizatorRequest RequestLocalizations(int languageIndex, LocalizatorRequestSettings settings,
+            string[][] localizations, Dictionary<string,int> stringMap, Action successfullyCompleded)
         {
-            string language = config.SupportedLanguages[languageIndex];
-            var request = new LocalizatorRequest(config, language, groups, localizations, successfullyCompleded);
-            if (!config.IsLanguageSupported(language))
+            string language = settings.config.SupportedLanguages[languageIndex];
+            var request = new LocalizatorRequest(settings, language, localizations, stringMap, successfullyCompleded);
+            if (!settings.config.IsLanguageSupported(language))
             {
                 request.Success = false;
                 request.Finished = true;
@@ -95,37 +96,76 @@ namespace AlchemyBow.Localizations
             using (var reader = new StringReader(localizationsFileContent))
             {
                 var csvReader = new CsvReader();
-                int columnIndex = config.DetermineFileColumnIndex(language);
-                int numberOfGroups = config.NumberOfGroups;
+                int columnIndex = settings.config.DetermineFileColumnIndex(language);
+                int numberOfGroups = settings.config.NumberOfGroups;
                 for (int i = 0; i < numberOfGroups; i++)
                 {
-                    int groupSize = config.GetGroupSize(i);
-                    if (groups == null || groups.Contains(i))
+                    if (settings.ShouldLoad(i))
                     {
-                        localizations[i] = new string[groupSize];
-                        for (int j = 0; j < groupSize; j++)
+                        if (settings.ShouldStringMap(i))
                         {
-                            string translation = csvReader.ReadRow(reader, columnIndex);
-                            if (translation != null)
-                            {
-                                localizations[i][j] = translation;
-                            }
-                            else
-                            {
-                                localizations[i][j] = $"NOT_TRANSLATED_[{i}][{j}]";
-                                Debug.LogWarning($"No translations for the key[{i}][{j}].");
-                            }
+                            LoadAndStringMapGroup(columnIndex, i, csvReader, reader);
+                        }
+                        else
+                        {
+                            LoadGroup(columnIndex, i, csvReader, reader);
                         }
                     }
                     else
                     {
-                        localizations[i] = null;
-                        for (int j = 0; j < groupSize; j++)
-                        {
-                            csvReader.ReadRow(reader, columnIndex);
-                        }
+                        SkipGroup(i, csvReader, reader);
                     }
                 }
+            }
+        }
+
+        private void LoadAndStringMapGroup(int columnIndex, int groupIndex, CsvReader csvReader, StringReader reader)
+        {
+            int groupSize = settings.config.GetGroupSize(groupIndex);
+            localizations[groupIndex] = new string[groupSize];
+            string groupName = settings.config.GetGroupName(groupIndex);
+            int groupFirstIndex = settings.config.GetGroupFirstIndex(groupIndex);
+            var row = new List<string>();
+            for (int j = 0; j < groupSize; j++)
+            {
+                row.Clear();
+                if (csvReader.ReadRow(row, reader) > columnIndex)
+                {
+                    localizations[groupIndex][j] = row[columnIndex];
+                    stringMap.Add($"{groupName}.{row[0]}", groupFirstIndex + j);
+                }
+                else
+                {
+                    localizations[groupIndex][j] = $"NOT_TRANSLATED_[{groupIndex}][{j}]";
+                    Debug.LogWarning($"No translations for the key[{groupIndex}][{j}].");
+                }
+            }
+        }
+        private void LoadGroup(int columnIndex, int groupIndex, CsvReader csvReader, StringReader reader)
+        {
+            int groupSize = settings.config.GetGroupSize(groupIndex);
+            localizations[groupIndex] = new string[groupSize];
+            for (int j = 0; j < groupSize; j++)
+            {
+                string translation = csvReader.ReadRow(reader, columnIndex);
+                if (translation != null)
+                {
+                    localizations[groupIndex][j] = translation;
+                }
+                else
+                {
+                    localizations[groupIndex][j] = $"NOT_TRANSLATED_[{groupIndex}][{j}]";
+                    Debug.LogWarning($"No translations for the key[{groupIndex}][{j}].");
+                }
+            }
+        }
+        private void SkipGroup(int groupIndex, CsvReader csvReader, StringReader reader)
+        {
+            int groupSize = settings.config.GetGroupSize(groupIndex);
+            localizations[groupIndex] = null;
+            for (int j = 0; j < groupSize; j++)
+            {
+                csvReader.ReadRow(reader, 0);
             }
         }
     }